+++ website/documentation-site.ts
@@ -21,6 +21,7 @@ export const DOCUMENTATION_DEFAULT_VERSION = 'latest';
 
 export const DOCUMENTATION_VERSIONS = [
   'latest',
+  '0.37.5',
   '0.37.4',
   '0.37.3',
   '0.37.2',